+++ assist/src/prepare.rs
@@ -251,7 +251,7 @@ impl Preparator {
         let space = self.config.data.account_size as u32;
         crate::common::iter_pdas(
             &self.keypairs,
-            self.config.payers as usize,
+            usize::from(self.config.payers),
             self.config.benchmark.accounts_count,
             space,
             self.config.authority,
@@ -270,7 +270,7 @@ impl Preparator {
 struct Pda {
     pubkey: Pubkey,
     payer: Keypair,
-    seed: u8,
+    seed: u16,
     bump: u8,
     space: u32,
 }
